# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "News aggregator API",
        "title": "Prospero",
        "contact": {
            "name": "Vitalii Popov",
            "url": "https://www.linkedin.com/in/mskkote/",
            "email": "msk.vitaly@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
        "/RSS/addSource": {
            "post": {
                "description": "Create a new RSS source",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Create new RSS source",
                "parameters": [
                    {
                        "description": "Add Source DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/source.AddSourceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/RSS/getEnrichedSources": {
            "get": {
                "description": "Read RSS sources with associated publishers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Read RSS sources with publishers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/RSS/getSources": {
            "get": {
                "description": "Read RSS sources with optional search and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Read RSS sources",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/RSS/removeSource": {
            "delete": {
                "description": "Delete RSS source by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Delete RSS source",
                "parameters": [
                    {
                        "description": "Delete Source DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/source.DeleteSourceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/RSS/updateSource": {
            "put": {
                "description": "Update existing RSS source",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Update RSS source",
                "parameters": [
                    {
                        "description": "Update Source DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/source.DTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/addPublisher": {
            "post": {
                "description": "Create a new publisher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publishers"
                ],
                "summary": "Create new publisher",
                "parameters": [
                    {
                        "description": "Add Publisher DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/publisher.AddPublisherDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/addSourceAndPublisher": {
            "post": {
                "description": "Add a new source along with its corresponding publisher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "Add Source and Publisher",
                "parameters": [
                    {
                        "description": "Add Source and Publisher DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/source.AddSourceAndPublisherDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/getPublishers": {
            "get": {
                "description": "Read publishers with optional search",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publishers"
                ],
                "summary": "Read publishers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/removePublisher": {
            "delete": {
                "description": "Delete publisher by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publishers"
                ],
                "summary": "Delete publisher",
                "parameters": [
                    {
                        "description": "Delete Publisher DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/publisher.DeletePublisherDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/updatePublisher": {
            "put": {
                "description": "Update publisher information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publishers"
                ],
                "summary": "Update publisher",
                "parameters": [
                    {
                        "description": "Publisher DTO",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/publisher.DTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "publisher.AddPublisherDTO": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "publisher.DTO": {
            "type": "object",
            "properties": {
                "add_date": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "publisher_id": {
                    "type": "string"
                }
            }
        },
        "publisher.DeletePublisherDTO": {
            "type": "object",
            "properties": {
                "publisher_id": {
                    "type": "string"
                }
            }
        },
        "source.AddSourceAndPublisherDTO": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "rss_Url": {
                    "type": "string"
                }
            }
        },
        "source.AddSourceDTO": {
            "type": "object",
            "properties": {
                "publisher_id": {
                    "type": "string"
                },
                "rss_url": {
                    "type": "string"
                }
            }
        },
        "source.DTO": {
            "type": "object",
            "properties": {
                "add_date": {
                    "type": "string"
                },
                "publisher_id": {
                    "type": "string"
                },
                "rss_id": {
                    "type": "string"
                },
                "rss_url": {
                    "type": "string"
                }
            }
        },
        "source.DeleteSourceDTO": {
            "type": "object",
            "properties": {
                "rss_id": {
                    "type": "string"
                }
            }
        }
    }
}